move metrics' spring config form server to core
