+++ integration/grpc/pom.xml
@@ -73,8 +73,10 @@
                 <artifactId>protobuf-maven-plugin</artifactId>
                 <configuration>
                     <protocArtifact>com.google.protobuf:protoc:3.3.0:exe:${os.detected.classifier}</protocArtifact>
+<!--                    <protocArtifact>com.google.protobuf:protoc:3.3.0:exe:osx-x86_64</protocArtifact>-->
                     <pluginId>grpc-java</pluginId>
                     <pluginArtifact>io.grpc:protoc-gen-grpc-java:1.8.0:exe:${os.detected.classifier}</pluginArtifact>
+<!--                    <pluginArtifact>io.grpc:protoc-gen-grpc-java:1.8.0:exe:osx-x86_64</pluginArtifact>-->
                 </configuration>
                 <executions>
                     <execution>
