+++ serializer/seata-serializer-protobuf/pom.xml
@@ -55,6 +55,7 @@
                     <protoSourceRoot>${project.basedir}/src/main/resources/protobuf/io/seata/protocol/transcation/</protoSourceRoot>
                     <protocArtifact>
                         com.google.protobuf:protoc:3.3.0:exe:${os.detected.classifier}
+<!--                        com.google.protobuf:protoc:3.3.0:exe:osx-x86_64-->
                     </protocArtifact>
                 </configuration>
                 <executions>
